Conflicts:
	src/game/AuctionHouseHandler.cpp
	src/game/AuctionHouseMgr.cpp
	src/game/Mail.cpp
	src/game/Player.cpp
	src/game/Player.h
	src/game/World.h

+++ src/game/PlayerbotMageAI.cpp
@@ -555,7 +555,7 @@ bool PlayerbotMageAI::DoNonCombatActions()
         Group::MemberSlotList const& groupSlot = pMaster->GetGroup()->GetMemberSlots();
         for (Group::member_citerator itr = groupSlot.begin(); itr != groupSlot.end(); itr++)
         {
-            Player *tPlayer = sObjectMgr.GetPlayer(uint64(itr->guid));
+            Player *tPlayer = sObjectMgr.GetPlayer(itr->guid);
             if (!tPlayer || !tPlayer->isAlive())
                 continue;
