+++ src/game/PlayerbotShamanAI.cpp
@@ -706,7 +706,7 @@ bool PlayerbotShamanAI::DoNonCombatActions()
         Group::MemberSlotList const& groupSlot = pMaster->GetGroup()->GetMemberSlots();
         for (Group::member_citerator itr = groupSlot.begin(); itr != groupSlot.end(); itr++)
         {
-            Player *tPlayer = sObjectMgr.GetPlayer(uint64(itr->guid));
+            Player *tPlayer = sObjectMgr.GetPlayer(itr->guid);
             if (!tPlayer || !tPlayer->isAlive())
                 continue;
